requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
google-api-core==2.11.0
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
